pkl-trans: fix 32-bit
2021-03-06  Jose E. Marchesi  <jemarch@gnu.org>

	* libpoke/pkl-trans.c (pkl_trans1_ps_array): Avoid using size_t
	variables, fixing 32-bit.

+++ libpoke/pkl-trans.c
@@ -988,8 +988,8 @@ PKL_PHASE_BEGIN_HANDLER (pkl_trans1_ps_array)
     {
       pkl_ast_node initializer_index_node
         = PKL_AST_ARRAY_INITIALIZER_INDEX (tmp);
-      size_t initializer_index;
-      size_t elems_appended, effective_index;
+      uint64_t initializer_index;
+      uint64_t elems_appended, effective_index;
 
       /* Set the index of the initializer.  */
       if (initializer_index_node == NULL)
